Add new vim modules and other changes
This is a step towards tidying up the toolkit and create a better
workflow environment. New vim modules were added and few old unused
modules were removed.

YouCompleteMe is now being used as a Debian package. It was easy to deal
with the huge dependency tree this way.

Various key re-maps and style changes were done in terminator
configuration. A new configuration was added for gtk3.0 to change
terminator's appearance. This might affect other GTK3 applications.